Basic attempt to modularize markup cleaning function
